pyproject.toml
[tool.poetry]
name = "osc-gitea-plugin"
version = "0.1.0"
description = "Plugin for osc to interact with gitea"
authors = ["Dan Čermák <dcermak@suse.com>"]
license = "GPL-2.0-or-later"
readme = "README.rst"
packages = [{include = "osc_gitea_plugin"}]

[tool.poetry.dependencies]
python = "^3.11"
osc = "^1.3"
py-gitea-opensuse-org = "^1.19"
pydantic = "<2.0.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
black = ">=23.7.0"
mypy = ">=1.7.0"
types-pyyaml = "^6.0"

[tool.poetry.scripts]
fork = "osc_gitea_plugin.fork_command:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
